fix a 64bit issue with a format string (remember that types that you don't
Thu Aug 23 23:23:42 2001  George Lebl <jirka@5z.com>

	* gnome-exec.c: fix a 64bit issue with a format string (remember
	  that types that you don't know size (such as gsize) should be
	  casted when passed to printf since they can be either %d or %ld
	  or some such.

+++ libgnome/gnome-exec.c
@@ -185,7 +185,8 @@ gnome_execute_async_with_env_fds (const char *dir, int argc,
   res = read (parent_comm_pipes[0], &child_pid, sizeof(child_pid));
   if (res != sizeof(child_pid))
     {
-      g_message("res is %d instead of %d", res, (int)sizeof(child_pid));
+      g_message("res is %ld instead of %d",
+		(long)res, (int)sizeof(child_pid));
       child_pid = -1; /* really weird things happened */
     }
   else if (read (parent_comm_pipes[0], &child_errno, sizeof(child_errno))
